feat(project): move data management to project
